pyproject.toml
[tool.black]
line-length = 80
target-version = [ "py310" ]

[tool.isort]
profile = "black"
line_length = 80
filter_files = true

[tool.poetry]
name = "backy"
version = "2.5.0dev"
description = ""
authors = [
    "Christian Theune <ct@flyingcircus.io>",
    "Christian Kauhaus <kc@flyingcircus.io>",
    "Daniel Kraft <daniel.kraft@d9t.de>"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: System :: Archiving :: Backup"
]
packages = [
    { include = "backy", from = "src" }
]
license = "GPLv3"

[tool.poetry.dependencies]
python = "~3.10"
PyYAML = "^5.3.1"
consulate-fc-nix-test = "1.1.0a1"
humanize = "^4.4.0"
mmh3 = "^3.0"
packaging = "^22.0"
prettytable = "^3.6.0"
python-lzo = "^1.14"
requests = "^2.28.0"
shortuuid = "^1.0.11"
structlog = "^22.3.0"
telnetlib3 = "^2.0.0"
tzlocal = "^4.2"
colorama = "^0.4.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.21.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
pytest-cache = "^1.0"
pytest-cov = "^4.0.0"
pytest-flake8 = "^1.1.1"
pytest-timeout = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
backy = "backy.main:main"

[tool.poetry.plugins."backy.sources"]
ceph-rbd = "backy.sources.ceph.source:CephRBD"
file = "backy.sources.file:File"
flyingcircus = "backy.sources.flyingcircus.source:FlyingCircusRootDisk"
